// packages/cv/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  blog: typeof blog;
  chat: typeof chat;
  chatnode: typeof chatnode;
  dbchat: typeof dbchat;
  durableapi: typeof durableapi;
  durablechat: typeof durablechat;
  file: typeof file;
  http: typeof http;
  httpstream: typeof httpstream;
  movie: typeof movie;
  ratelimit: typeof ratelimit;
  retry: typeof retry;
  streaming: typeof streaming;
  "tools/weather": typeof tools_weather;
  toolschemas: typeof toolschemas;
  user: typeof user;
  workpool: typeof workpool;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;